+++ jGALiWeather/src/main/java/jgaliweather/configuration/partition_reader/Set.java
@@ -7,5 +7,5 @@ public interface Set {
     @Override
     public String toString();
     
-    public int apply(double value);
+    public double apply(double value);
 }
